Move tests for generated files into generated file test folder
Tests for generator belong in `/generator/tests`, tests for generated files (and manual code in `/lib`) belong in `/tests`

This PR is a little trickier than simply moving files because a couple of the tests had conflicting filenames (ie, `tests/GeneratedFilesTest.php` and `generator/tests/GeneratedFilesTests.php` both exist, with different contents, so they needed to be manually merged)
